using stable_baselines more heavily now

+++ src/env.py
@@ -28,4 +28,4 @@ class ShowdownEnv(Gen4EnvSinglePlayer[npt.NDArray[np.float32], int]):
         return np.array(Agent.embed_battle(battle))
 
     def describe_embedding(self) -> Space[npt.NDArray[np.float32]]:
-        return Box(0.0, 1.0, shape=(1404,))
+        return Box(0.0, 1.0, shape=(1404,), dtype=np.float32)
